Fix handling of missing numeric values
Unknown loan and transfer fees (usually denoted by '-', '?', or a blank space) are now treated as null values instead of as 0. Added a catch for missing ages and player IDs in the same way. Patched the parsing of club names from images by getting the alt text instead of the title text.